perf: lowered memory footprint of docker images
docker images now use alpine linux

+++ server/scripts/syncTables.js
@@ -1,5 +1,5 @@
-/* WARNING: On running this file all existing 
-tables defined in the models will be dropped */
+/* WARNING: On running this file with the -f flag, all existing
+tables with names defined in the models will be dropped from the db */
 
 const db = require('../config/database')
 const Metabolite = require('../models/metabolite')
@@ -7,4 +7,7 @@ const Model = require('../models/model')
 const Reaction = require('../models/reaction')
 const sbmlFile = require('../models/sbmlFile')
 
-db.sync({ force: true })
+if (process.argv[2] === '-f') {
+  console.warn('Force option has been used.')
+  db.sync({ force: true })
+} else db.sync()
